pyproject.toml
[tool.poetry]
name = "honu_near_agent_spike"
version = "0.1.0"
description = "Test Stack for a Honu x Near Agent"
authors = ["freyamade <freya.broderick@honu.ai>"]
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = "^0.38.0"
pydantic-settings = "^2.1.0"
structlog = "^25.4.0"
google-genai = "^1.3.0"
honu-google-adk = {git = "https://github.com/honu-ai/honu-google-adk.git", rev = "main"}
pillow = "^12.1.0"
google-adk = "1.16"
psycopg2-binary = "^2.9.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
